Feat/replace double for bigdecimal

+++ src/main/kotlin/bass/dto/PaymentDTO.kt
@@ -1,11 +1,12 @@
 package bass.dto
 
 import bass.entities.PaymentEntity
+import java.math.BigDecimal
 
 class PaymentDTO(
     val id: Long? = null,
     val stripePaymentIntentId: String,
-    val amount: Long,
+    val amount: BigDecimal,
     val currency: String,
     val status: PaymentEntity.PaymentStatus,
     val failureCode: String? = null,
